pyproject.toml
[tool.poetry]
name = "laboratorium-ai-asr"
version = "1.0.2"
description = ""
authors = ["Mathias Eulers <m.eulers@uni-luebeck.de>", "Finn Stoldt <f.stoldt@uni-luebeck.de>"]
readme = "README.md"
package-mode = true
packages = [{include = "laboratorium_ai_asr"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
librosa = "^0.10.1"
whisperx = "^3.1.3"
ctranslate2 = "^4.2.1"
openai-whisper = {git = "https://github.com/openai/whisper.git"}

[tool.poetry.scripts]
main = "laboratorium_ai_asr.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"